+++ src/gotenberg_client/_chromium/routes.py
@@ -11,11 +11,15 @@ from gotenberg_client._chromium.mixins import CookiesMixin
 from gotenberg_client._chromium.mixins import CssPageSizeMixin
 from gotenberg_client._chromium.mixins import CustomHTTPHeaderMixin
 from gotenberg_client._chromium.mixins import DocumentOutlineMixin
+from gotenberg_client._chromium.mixins import EmulatedMediaFeaturesMixin
 from gotenberg_client._chromium.mixins import EmulatedMediaMixin
+from gotenberg_client._chromium.mixins import GenerateTaggedPdfMixin
 from gotenberg_client._chromium.mixins import HeaderFooterMixin
+from gotenberg_client._chromium.mixins import IgnoreResourceDomainsMixin
 from gotenberg_client._chromium.mixins import InvalidStatusCodesMixin
 from gotenberg_client._chromium.mixins import MarginMixin
 from gotenberg_client._chromium.mixins import NativePageRangeMixin
+from gotenberg_client._chromium.mixins import NetworkAlmostIdleMixin
 from gotenberg_client._chromium.mixins import NetworkErrorsMixin
 from gotenberg_client._chromium.mixins import OmitBackgroundMixin
 from gotenberg_client._chromium.mixins import PageOrientMixin
@@ -23,9 +27,11 @@ from gotenberg_client._chromium.mixins import PageSizeMixin
 from gotenberg_client._chromium.mixins import PerformanceModeMixin
 from gotenberg_client._chromium.mixins import PrintBackgroundMixin
 from gotenberg_client._chromium.mixins import RenderControlMixin
+from gotenberg_client._chromium.mixins import ResourceStatusCodesMixin
 from gotenberg_client._chromium.mixins import ScaleMixin
 from gotenberg_client._chromium.mixins import ScreenShotSettingsMixin
 from gotenberg_client._chromium.mixins import SinglePageMixin
+from gotenberg_client._chromium.mixins import WaitForSelectorMixin
 from gotenberg_client._common import DownloadFromMixin
 from gotenberg_client._common import EmbedsMixin
 from gotenberg_client._common import EncryptMixin
@@ -54,13 +60,19 @@ class _BaseChromiumConvertMixin(
     NativePageRangeMixin,
     HeaderFooterMixin,
     RenderControlMixin,
+    WaitForSelectorMixin,
     EmulatedMediaMixin,
+    EmulatedMediaFeaturesMixin,
     CookiesMixin,
     CustomHTTPHeaderMixin,
     InvalidStatusCodesMixin,
+    ResourceStatusCodesMixin,
+    IgnoreResourceDomainsMixin,
     NetworkErrorsMixin,
     ConsoleExceptionMixin,
     PerformanceModeMixin,
+    NetworkAlmostIdleMixin,
+    GenerateTaggedPdfMixin,
     SplitModeMixin,
     PdfFormatMixin,
     PdfUniversalAccessMixin,
@@ -364,12 +376,17 @@ class AsyncMarkdownToPdfRoute(_BaseMarkdownToPdfRoute, AsyncBaseRoute):
 
 class _BaseScreenShotSettingsMixin(
     RenderControlMixin,
+    WaitForSelectorMixin,
     EmulatedMediaMixin,
+    EmulatedMediaFeaturesMixin,
     CookiesMixin,
     CustomHTTPHeaderMixin,
     InvalidStatusCodesMixin,
+    ResourceStatusCodesMixin,
+    IgnoreResourceDomainsMixin,
     ConsoleExceptionMixin,
     PerformanceModeMixin,
+    NetworkAlmostIdleMixin,
     OmitBackgroundMixin,
     ScreenShotSettingsMixin,
 ):