+++ src/gotenberg_client/_libreoffice/routes.py
@@ -20,12 +20,16 @@ from gotenberg_client._common import StampMixin
 from gotenberg_client._common import WatermarkMixin
 from gotenberg_client._libreoffice.mixins import LibreOfficeCompressOptionsMixin
 from gotenberg_client._libreoffice.mixins import LibreOfficeMergeOptionMixin
+from gotenberg_client._libreoffice.mixins import LibreOfficeNativeWatermarkMixin
 from gotenberg_client._libreoffice.mixins import LibreOfficePagePropertiesMixin
+from gotenberg_client._libreoffice.mixins import LibreOfficeViewerPreferencesMixin
 from gotenberg_client._typing_compat import Self
 
 
 class _BaseOfficeDocumentToPdfRoute(
     LibreOfficePagePropertiesMixin,
+    LibreOfficeNativeWatermarkMixin,
+    LibreOfficeViewerPreferencesMixin,
     LibreOfficeCompressOptionsMixin,
     LibreOfficeMergeOptionMixin,
     SplitModeMixin,