+++ tests/test_convert_pdf_a.py
@@ -1,11 +1,11 @@
 # SPDX-FileCopyrightText: 2023-present Trenton H <rda0128ou@mozmail.com>
 #
 # SPDX-License-Identifier: MPL-2.0
+from http import HTTPStatus
 from pathlib import Path
 
 import pikepdf
 import pytest
-from httpx import codes
 
 from gotenberg_client import GotenbergClient
 from gotenberg_client._pdfa_ua.routes import AsyncConvertToArchiveFormatRoute
@@ -28,7 +28,7 @@ class TestPdfAConvert:
         with sync_client.pdf_convert.to_pdfa() as route:
             resp = route.convert(pdf_sample_one_file).pdf_format(gt_format).run_with_retry()
 
-        assert resp.status_code == codes.OK
+        assert resp.status_code == HTTPStatus.OK
         assert "Content-Type" in resp.headers
         assert resp.headers["Content-Type"] == "application/pdf"
 
@@ -51,7 +51,7 @@ class TestPdfAConvert:
         with sync_client.pdf_convert.to_pdfa() as route:
             resp = route.convert_files([pdf_sample_one_file, other_test_file]).pdf_format(gt_format).run_with_retry()
 
-            assert resp.status_code == codes.OK
+            assert resp.status_code == HTTPStatus.OK
             assert "Content-Type" in resp.headers
             assert resp.headers["Content-Type"] == "application/zip"
 
@@ -65,7 +65,7 @@ class TestPdfAConvert:
                 route.convert(pdf_sample_one_file).pdf_format(PdfAFormat.A2b).enable_universal_access().run_with_retry()
             )
 
-        assert resp.status_code == codes.OK
+        assert resp.status_code == HTTPStatus.OK
         assert "Content-Type" in resp.headers
         assert resp.headers["Content-Type"] == "application/pdf"
 
@@ -82,7 +82,7 @@ class TestPdfAConvert:
                 .run_with_retry()
             )
 
-        assert resp.status_code == codes.OK
+        assert resp.status_code == HTTPStatus.OK
         assert "Content-Type" in resp.headers
         assert resp.headers["Content-Type"] == "application/pdf"
 
@@ -100,6 +100,6 @@ class TestPdfAConvertAsync:
             .run_with_retry()
         )
 
-        assert resp.status_code == codes.OK
+        assert resp.status_code == HTTPStatus.OK
         assert "Content-Type" in resp.headers
         assert resp.headers["Content-Type"] == "application/pdf"