+++ tests/test_merge.py
@@ -1,11 +1,11 @@
 # SPDX-FileCopyrightText: 2023-present Trenton H <rda0128ou@mozmail.com>
 #
 # SPDX-License-Identifier: MPL-2.0
+from http import HTTPStatus
 from pathlib import Path
 
 import pikepdf
 import pytest
-from httpx import codes
 from pytest_httpx import HTTPXMock
 
 from gotenberg_client import GotenbergClient
@@ -37,7 +37,7 @@ class TestMergePdfs:
                 )
                 .run_with_retry()
             )
-        assert resp.status_code == codes.OK
+        assert resp.status_code == HTTPStatus.OK
         assert "Content-Type" in resp.headers
         assert resp.headers["Content-Type"] == "application/pdf"
 
@@ -59,7 +59,7 @@ class TestMergePdfs:
                 [sample_directory / "z_first_merge.pdf", sample_directory / "a_merge_second.pdf"],
             ).run_with_retry()
 
-            assert resp.status_code == codes.OK
+            assert resp.status_code == HTTPStatus.OK
             assert "Content-Type" in resp.headers
             assert resp.headers["Content-Type"] == "application/pdf"
 
@@ -85,7 +85,7 @@ class TestMergePdfsAsync:
             [sample_directory / "z_first_merge.pdf", sample_directory / "a_merge_second.pdf"],
         ).run_with_retry()
 
-        assert resp.status_code == codes.OK
+        assert resp.status_code == HTTPStatus.OK
         assert "Content-Type" in resp.headers
         assert resp.headers["Content-Type"] == "application/pdf"
 
@@ -121,3 +121,26 @@ class TestMergePdfsMocked:
         with sync_client.merge.merge() as route:
             route.merge([sample_directory / "sample1.pdf"]).user_password("pw").run()
         verify_stream_contains(httpx_mock.get_request(), "userPassword", "pw")
+
+    def test_merge_auto_index_bookmarks(
+        self,
+        sync_client: GotenbergClient,
+        sample_directory: Path,
+        httpx_mock: HTTPXMock,
+    ):
+        httpx_mock.add_response(method="POST")
+        with sync_client.merge.merge() as route:
+            route.merge([sample_directory / "sample1.pdf"]).auto_index_bookmarks(enable=True).run()
+        verify_stream_contains(httpx_mock.get_request(), "autoIndexBookmarks", "true")
+
+    def test_merge_bookmarks(
+        self,
+        sync_client: GotenbergClient,
+        sample_directory: Path,
+        httpx_mock: HTTPXMock,
+    ):
+        httpx_mock.add_response(method="POST")
+        bookmarks = [{"title": "Section 1", "page": 1}]
+        with sync_client.merge.merge() as route:
+            route.merge([sample_directory / "sample1.pdf"]).merge_bookmarks(bookmarks).run()
+        verify_stream_contains(httpx_mock.get_request(), "bookmarks", "Section 1")